[minor] Move delete log get to a separate file
